[MECHANICS_PLAYGROUND] Create view models for win and death screens
