+++ app/src/main/java/ru/meatgames/tomb/screen/compose/death/DeathScreen.kt
@@ -1,4 +1,4 @@
-package ru.meatgames.tomb.screen.compose
+package ru.meatgames.tomb.screen.compose.death
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
@@ -11,18 +11,19 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import androidx.hilt.navigation.compose.hiltViewModel
 import ru.meatgames.tomb.design.component.BaseTextButton
 import ru.meatgames.tomb.design.h1TextStyle
 
-@Preview(widthDp = 360, heightDp = 640,)
+@Preview(widthDp = 360, heightDp = 640)
 @Composable
-fun WinScreenPreview() {
-    WinScreen { Unit }
+fun DeathScreenPreview() {
+    DeathScreen()
 }
 
 @Composable
-fun WinScreen(
-    onNavigateToMainMenu: () -> Unit,
+fun DeathScreen(
+    viewModel: DeathScreenVM = hiltViewModel(),
 ) {
     Box(
         modifier = Modifier
@@ -30,17 +31,17 @@ fun WinScreen(
             .background(Color(0xFF212121)),
     ) {
         Text(
-            text = "You won!",
+            text = "You are dead...",
             modifier = Modifier.align(Alignment.Center),
             style = h1TextStyle,
         )
-
+    
         BaseTextButton(
             title = "To main menu",
             modifier = Modifier
                 .align(Alignment.BottomCenter)
                 .padding(bottom = 16.dp),
-            onClick = onNavigateToMainMenu,
+            onClick = viewModel::navigateToMainMenu,
         )
     }
 }
